Moved the configuration read and zone related pages
