+++ src/lib/y2firewall/widgets/allowed_services.rb
@@ -65,6 +65,7 @@ module Y2Firewall
       end
 
     private
+
       # @!attribute [r] available_svcs_table
       #   @return [ServicesTable]
       #
