+++ src/lib/y2firewall/widgets/pages/interfaces.rb
@@ -33,8 +33,6 @@ module Y2Firewall
         # @param pager [CWM::TreePager]
         def initialize(_pager)
           textdomain "firewall"
-          @fw = Y2Firewall::Firewalld.instance
-          @fw.read # FIXME: when?
         end
 
         # @macro seeAbstractWidget
